requirements.txt
allure-pytest==2.2.3b1
allure-python-commons==2.2.3b1
Appium-Python-Client==0.24
attrs==17.2.0
certifi==2017.7.27.1
chardet==3.0.4
enum34==1.1.6
flaky==3.4.0
idna==2.6
jsonschema==2.6.0
lxml==4.1.0
namedlist==1.7
pluggy==0.5.2
py==1.4.34
pytest==3.2.3
requests==2.18.4
selenium==3.7.0
six==1.11.0
urllib3==1.22
